jarvis_desktop/requirements.txt
PyQt5==5.15.9
PyQt5-sip==12.13.0
SpeechRecognition==3.10.1
pyttsx3==2.90
python-dotenv==1.0.0
anthropic>=0.7.0
requests==2.31.0
pywin32==306
python-vlc==3.0.20125
spotipy==2.22.1
pynput==1.7.6
pillow==10.0.0
